+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 **?.?.? Changes:**
 
-* Added 4 new effects:
+* Added 5 new effects:
   * Activate All Interactables: Interacts with most interactable objects on the stage (chests, shrines)
   * Explosive Deaths: All non-players explode at low health or death, lasts 90 seconds
   * Prototype Models: Replaces all character models with their hitboxes
   * Pause: Pauses the run timer and effect activation, lasts 60 seconds
+  * 2x Pickups: Doubles all dropped pickups, lasts 90 seconds
 
 * Renamed 1 effect:
   * Extreme Recoil -> Knockback Recoil